requirements-gcp.txt
# HIMARI Layer 3 - GCP Deployment Requirements

# Core ML
torch==2.0.1
numpy==1.24.3

# GCP SDKs
google-cloud-aiplatform==1.38.0
google-cloud-storage==2.10.0
google-cloud-monitoring==2.15.1
google-cloud-logging==3.5.0

# API Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# HTTP client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
